fixed NPE which occurs while shutdown
왜인지는 아직 잘 모르겠지만 스위퍼가 제때 종료되지 않는 것으로 보임. 다른 곳에서 온라인 라이터들을 플러시/정리 하고 있는 와중에 스위퍼가 돌 가능성은 있다고 보여서, 해당 동작을 합법적인 것으로 처리하기로 함. 별도로, stop() 에서 스위퍼 정리 순서가 잘못되어 있어서 수정함.

+++ araqne-logstorage/src/main/java/org/araqne/logstorage/engine/LogStorageEngine.java
@@ -243,6 +243,11 @@ public class LogStorageEngine implements LogStorage, TableEventListener, LogFile
 				throw e;
 		}
 
+		writerSweeper.doStop = true;
+		synchronized (writerSweeper) {
+			writerSweeper.notifyAll();
+		}
+
 		// wait writer sweeper stop
 		try {
 			for (int i = 0; i < 25; i++) {
@@ -263,11 +268,6 @@ public class LogStorageEngine implements LogStorage, TableEventListener, LogFile
 			} catch (Throwable t) {
 				logger.warn("exception caught", t);
 			}
-		}
-
-		writerSweeper.doStop = true;
-		synchronized (writerSweeper) {
-			writerSweeper.notifyAll();
 		}
 
 		onlineWriters.clear();
@@ -995,6 +995,8 @@ public class LogStorageEngine implements LogStorage, TableEventListener, LogFile
 				this.flushAll = false;
 				for (OnlineWriterKey key : onlineWriters.keySet()) {
 					OnlineWriter writer = onlineWriters.get(key);
+					if (writer == null)
+						continue;
 					boolean doFlush = writer.isCloseReserved() || ((now - writer.getLastFlush().getTime()) > flushInterval);
 					doFlush = flushAll ? true : doFlush;
 					if (doFlush) {
